backend/requirements.txt
cloudinary==1.36.0
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sendgrid==6.10.0
algoliasearch==3.0.0
flask-swagger-ui==4.11.1